Move Formatting from AshRefCommand into KoLmafiaASH
Fixes the inverted dependency where KoLmafiaASH depended on a command
class. Now both ref commands and KoLmafiaASH.showUserFunctions use
KoLmafiaASH.Formatting. Also extract magic color strings to named
constants.
